chore: Refactor for best practices
Closes #26
