+++ src/app/core/services/battle-net/battle-net.interface.ts
