+++ src/app/core/services/battle-net/battle-net.service.ts
