+++ src/app/core/services/battle-net/character/character-image.pipe.ts
