+++ src/app/core/services/battle-net/character/character.interface.ts
