+++ src/app/core/services/battle-net/character/character.service.ts
