+++ src/app/core/services/battle-net/character/types/battlenet-general.ts
