+++ src/app/core/services/battle-net/character/types/battlenet-media.ts
