+++ src/app/core/services/character-store/character-store.interface.ts
