+++ src/app/core/services/character-store/character-store.service.ts
