+++ src/app/core/services/checklist-evaluator/handlers/_handler.interface.ts
