+++ src/app/core/services/checklist/checklist.interface.ts
