+++ src/app/core/services/responsive/responsive.service.ts
