+++ src/app/pages/characters/characters.component.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .card-container {
     text-align: left;
