+++ src/app/pages/checklist/checklist-line/checklist-line.component.scss
