+++ src/app/pages/checklist/checklist.component.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .page-actions, mat-card {
     width: 100%;
