+++ src/app/shared/material-common.module.ts
@@ -1,0 +1,32 @@
+import { NgModule } from '@angular/core';
+import {
+    MatButtonModule,
+    MatCardModule,
+    MatCheckboxModule,
+    MatIconModule,
+    MatProgressSpinnerModule,
+    MatSidenavModule,
+    MatToolbarModule,
+} from '@angular/material';
+
+@NgModule({
+    imports: [
+        MatToolbarModule,
+        MatSidenavModule,
+        MatIconModule,
+        MatButtonModule,
+        MatCardModule,
+        MatCheckboxModule,
+        MatProgressSpinnerModule,
+    ],
+    exports: [
+        MatToolbarModule,
+        MatSidenavModule,
+        MatIconModule,
+        MatButtonModule,
+        MatCardModule,
+        MatCheckboxModule,
+        MatProgressSpinnerModule,
+    ],
+})
+export class MaterialCommonModule { }
