+++ src/app/shared/util/object-util.ts
